basic logic for distinct views and actions per user role

+++ CinemaApplication/Controllers/ScreeningController.cs
@@ -2,7 +2,7 @@
 using CinemaApplication.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-
+using Microsoft.AspNetCore.Http;
 namespace CinemaApplication.Controllers
 {
     public class ScreeningController : Controller
@@ -21,15 +21,20 @@ namespace CinemaApplication.Controllers
         public IActionResult Index(String MovieName)
         {
             TempData["MovieName"] = MovieName;
+            TempData["role"] = HttpContext.Session.GetString("role");
             IEnumerable<Screening> movieScreenings = _db.Screenings.Where(s => s.Movie.MovieName == MovieName && s.Time>DateTime.Now).Include(s=> s.Cinema);
             return View(movieScreenings);
         }
         [HttpPost]
         public IActionResult Create(String MovieName,String CinemaName,DateTime Time)
         {
+            if (!HttpContext.Session.GetString("role").Equals("CONTENT-ADMIN"))
+            {
+                return RedirectToAction("Index", "Movie");
+            }
             Cinema cinema = _db.Cinemas.FirstOrDefault(c=>c.Name == CinemaName);
             Movie movie = _db.Movies.FirstOrDefault(m=>m.MovieName==MovieName);
-            ContentAdmin contentAdmin = _db.Set<ContentAdmin>().FirstOrDefault(u => u.Username.Equals("cadmin"));
+            ContentAdmin contentAdmin = _db.Set<ContentAdmin>().FirstOrDefault(u => u.Username.Equals(HttpContext.Session.GetString("username")));
             Screening screening = new Screening();
             screening.Movie = movie;
             screening.ContentAdmin = contentAdmin;
@@ -43,6 +48,10 @@ namespace CinemaApplication.Controllers
         [HttpGet]
         public IActionResult Edit(int? Id)
         {
+            if (!HttpContext.Session.GetString("role").Equals("CONTENT-ADMIN"))
+            {
+                return RedirectToAction("Index", "Movie");
+            }
             Screening screening = _db.Screenings.Include(s => s.Movie).Include(s => s.Cinema).FirstOrDefault(s => s.Id == Id);
             return View(screening);
         }
@@ -51,14 +60,21 @@ namespace CinemaApplication.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult Edit(Screening screening,String MovieName)
         {
+            if (!HttpContext.Session.GetString("role").Equals("CONTENT-ADMIN"))
+            {
+                return RedirectToAction("Index", "Movie");
+            }
             _db.Entry(screening).State = EntityState.Modified;
             _db.SaveChanges();
             return RedirectToAction("Index", "Screening",new {MovieName=MovieName});
-           
         }
 
         public IActionResult Delete(int? Id)
         {
+            if (!HttpContext.Session.GetString("role").Equals("CONTENT-ADMIN"))
+            {
+                return RedirectToAction("Index", "Movie");
+            }
             Screening screening = _db.Screenings.Include(s=>s.Movie).FirstOrDefault(s=> s.Id==Id);
             String movieName = screening.Movie.MovieName;
             _db.Screenings.Remove(screening);
